Changed icon-set. Updated DebugConsole.cs WriteLine's outut format (Issue70). Fixed Setup Alarm bug (Issue74). Plugins do not anymore try to update again while an update already runs.

+++ LibEvents/frmSetupAlarm.cs
@@ -27,7 +27,7 @@ namespace LibEvents
 
             foreach (byte m in minutes)
             {
-                if (m < (byte)this._event.MinutesLeft) this.ComboBoxAlertBefore.Items.Add(m);
+                if ((double)m < this._event.MinutesLeft) this.ComboBoxAlertBefore.Items.Add(m);
             }
 
             if (this.ComboBoxAlertBefore.Items.Count > 0) this.ComboBoxAlertBefore.SelectedIndex = 0;
